Autocomplete: Ship jaccard similarity changes
