+++ vscode/src/completions/completion-provider-config.ts
@@ -11,7 +11,6 @@ class CompletionProviderConfig {
 
     private flagsToResolve = [
         FeatureFlag.CodyAutocompleteContextBfgMixed,
-        FeatureFlag.CodyAutocompleteContextNewJaccardSimilarity,
         FeatureFlag.CodyAutocompleteDynamicMultilineCompletions,
         FeatureFlag.CodyAutocompleteHotStreak,
         FeatureFlag.CodyAutocompleteSingleMultilineRequest,
@@ -77,9 +76,6 @@ class CompletionProviderConfig {
         const { config } = this
 
         const bfgMixedContextFlag = this.getPrefetchedFlag(FeatureFlag.CodyAutocompleteContextBfgMixed)
-        const newJaccardSimilarityContextFlag = this.getPrefetchedFlag(
-            FeatureFlag.CodyAutocompleteContextNewJaccardSimilarity
-        )
 
         const contextStrategy: ContextStrategy =
             config.autocompleteExperimentalGraphContext === 'bfg'
@@ -94,9 +90,7 @@ class CompletionProviderConfig {
                               ? 'new-jaccard-similarity'
                               : bfgMixedContextFlag
                                   ? 'bfg-mixed'
-                                  : newJaccardSimilarityContextFlag
-                                      ? 'new-jaccard-similarity'
-                                      : 'jaccard-similarity'
+                                  : 'jaccard-similarity'
 
         return contextStrategy
     }
