+++ vscode/src/completions/context/retrievers/jaccard-similarity/bestJaccardMatch.test.ts
@@ -1,6 +1,7 @@
+import dedent from 'dedent'
 import { describe, expect, it } from 'vitest'
 
-import { bestJaccardMatch, getWords } from './bestJaccardMatch'
+import { bestJaccardMatches, getWordOccurrences } from './bestJaccardMatch'
 
 const targetSnippet = `
 import { bestJaccardMatch, getWords } from './context'
@@ -25,30 +26,18 @@ describe('getWords', () => {
 })
 `
 
-const matchSnippet = `
-describe('bestJaccardMatch', () => {
-    it('should return the best match', () => {
-        const matchText = [
-            'foo',
-            'bar',
-            'baz',
-            'qux',
-            'quux',
-        ].join('\n')
-    })
-})
-`
+const MAX_MATCHES = 50
 
 describe('getWords', () => {
     it('works with regular text', () => {
-        expect(getWords('foo bar baz')).toEqual(
+        expect(getWordOccurrences('foo bar baz')).toEqual(
             new Map<string, number>([
                 ['foo', 1],
                 ['bar', 1],
                 ['baz', 1],
             ])
         )
-        expect(getWords('running rocks slipped over')).toEqual(
+        expect(getWordOccurrences('running rocks slipped over')).toEqual(
             new Map<string, number>([
                 ['run', 1],
                 ['rock', 1],
@@ -58,7 +47,7 @@ describe('getWords', () => {
     })
 
     it('works with code snippets', () => {
-        expect(getWords(targetSnippet)).toEqual(
+        expect(getWordOccurrences(targetSnippet)).toEqual(
             new Map<string, number>([
                 ['import', 1],
                 ['bestjaccardmatch', 1],
@@ -87,61 +76,158 @@ describe('getWords', () => {
 })
 
 describe('bestJaccardMatch', () => {
-    it('should return the best match in 5 line windows', () => {
-        const matchText = [
-            'foo',
-            'bar',
-            'baz',
-            'qux',
-            'quux',
-            'quuz',
-            'corge',
-            'grault',
-            'garply',
-            'waldo',
-            'fred',
-            'plugh',
-            'xyzzy',
-            'thud',
-        ].join('\n')
-        expect(bestJaccardMatch('foo\nbar\nbaz', matchText, 3)).toEqual({
+    it('should return the best match', () => {
+        const matchText = dedent`
+            foo
+            bar
+            baz
+            qux
+            quux
+            quuz
+            corge
+            grault
+            garply
+            waldo
+            fred
+            plugh
+            xyzzy
+            thud
+        `
+        expect(bestJaccardMatches('foo\nbar\nbaz', matchText, 3, MAX_MATCHES)[0]).toEqual({
             score: 1,
             content: 'foo\nbar\nbaz',
-            startLine: 0,
             endLine: 2,
+            startLine: 0,
         })
-        expect(bestJaccardMatch('bar\nquux', matchText, 4)).toEqual({
+        expect(bestJaccardMatches('bar\nquux', matchText, 4, MAX_MATCHES)[0]).toEqual({
             score: 0.5,
             content: 'bar\nbaz\nqux\nquux',
-            startLine: 1,
             endLine: 4,
+            startLine: 1,
         })
         expect(
-            bestJaccardMatch(
+            bestJaccardMatches(
                 ['grault', 'notexist', 'garply', 'notexist', 'waldo', 'notexist', 'notexist'].join('\n'),
                 matchText,
-                6
-            )
+                6,
+                MAX_MATCHES
+            )[0]
         ).toEqual({
             score: 0.3,
-            content: ['corge', 'grault', 'garply', 'waldo', 'fred', 'plugh'].join('\n'),
-            startLine: 6,
-            endLine: 11,
+            startLine: 4,
+            endLine: 9,
+            content: ['quux', 'quuz', 'corge', 'grault', 'garply', 'waldo'].join('\n'),
         })
     })
 
+    it('returns more than one match', () => {
+        const matchText = dedent`
+            foo
+            bar
+            baz
+            qux
+            foo
+            quuz
+            corge
+            grault
+            garply
+            waldo
+            fred
+            plugh
+            xyzzy
+            thud`
+
+        const matches = bestJaccardMatches('foo\nbar\nbaz', matchText, 3, MAX_MATCHES)
+
+        expect(matches).toHaveLength(4)
+        expect(matches.map(match => match.content.split('\n'))).toEqual([
+            ['foo', 'bar', 'baz'],
+            ['qux', 'foo', 'quuz'],
+            ['corge', 'grault', 'garply'],
+            ['waldo', 'fred', 'plugh'],
+        ])
+    })
+
     it('works with code snippets', () => {
-        expect(bestJaccardMatch(targetSnippet, matchSnippet, 5)).toMatchInlineSnapshot(`
+        expect(
+            bestJaccardMatches(
+                targetSnippet,
+                dedent`
+                    describe('bestJaccardMatch', () => {
+                        it('should return the best match', () => {
+                            const matchText = [
+                                'foo',
+                                'bar',
+                                'baz',
+                                'qux',
+                                'quux',
+                            ].join('\n')
+                        })
+                    })
+                `,
+                5,
+                MAX_MATCHES
+            )[0]
+        ).toMatchInlineSnapshot(`
           {
             "content": "describe('bestJaccardMatch', () => {
               it('should return the best match', () => {
                   const matchText = [
                       'foo',
                       'bar',",
-            "endLine": 5,
+            "endLine": 4,
             "score": 0.08695652173913043,
-            "startLine": 1,
+            "startLine": 0,
           }
         `)
+    })
+
+    it('works for input texts that are shorter than the window size', () => {
+        expect(bestJaccardMatches('foo', 'foo', 10, MAX_MATCHES)[0]).toEqual({
+            content: 'foo',
+            endLine: 0,
+            score: 1,
+            startLine: 0,
+        })
+    })
+
+    it('skips over windows with empty start lines', () => {
+        const matches = bestJaccardMatches(
+            'foo',
+            dedent`
+                // foo
+                // unrelated 1
+                // unrelated 2
+
+
+                // foo
+                // unrelated 3
+                // unrelated 4
+            `,
+            3,
+            MAX_MATCHES
+        )
+
+        expect(matches[0].content).toBe('// foo\n// unrelated 1\n// unrelated 2')
+        expect(matches[1].content).toBe('// foo\n// unrelated 3\n// unrelated 4')
+    })
+
+    it("does not skips over windows with empty start lines if we're at the en", () => {
+        const matches = bestJaccardMatches(
+            targetSnippet,
+            dedent`
+                // foo
+                // unrelated
+                // unrelated
+
+
+                // foo
+            `,
+            3,
+            MAX_MATCHES
+        )
+
+        expect(matches[0].content).toBe('\n\n// foo')
+        expect(matches[1].content).toBe('// foo\n// unrelated\n// unrelated')
     })
 })
